Update draft-ietf-ccamp-mw-topo-yang.txt
Some editorial updates.
How about the text in the revision statement. Still valid?
Suggesting to remove the typedef for power. Only used once.
tree: this version is truncated at the end, but geolocation to be moved to node, ifref- leafs to be removed - Will fix that separately
Have not reviewed the JSON instantiations. Will do that once they have been updated